acpi: acpidump: fix overflow before widening bug
Coverity CID#997275, Unintentional integer overflow (OVERFLOW_BEFORE_WIDEN)

To avoid 32 bit overflow, cast info->bit_field_nbits to uint64_t before the
shift.

Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>

+++ src/acpi/acpidump/acpidump.c
@@ -101,7 +101,7 @@ static void acpi_dump_str(fwts_framework *fw, fwts_acpidump_field *info, void *d
 static uint64_t apci_dump_get_uint64_t(fwts_acpidump_field *info, void *data)
 {
 	uint64_t	ret;
-	uint64_t	mask = (1<<info->bit_field_nbits) - 1;
+	uint64_t	mask = (1<<(uint64_t)info->bit_field_nbits) - 1;
 	int i;
 
 	switch (info->size) {
